use set in recoil selector
