+++ src/App.tsx
@@ -1,86 +1,21 @@
-import styled, { createGlobalStyle } from "styled-components";
-import ToDoList from "./components/ToDoList";
-
-const GlobalStyle = createGlobalStyle`
-	/* http://meyerweb.com/eric/tools/css/reset/
-   v5.0.1 | 20191019
-   License: none (public domain)
-*/
-
-@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100..900');
-
-html, body, div, span, applet, object, iframe,
-h1, h2, h3, h4, h5, h6, p, blockquote, pre,
-a, abbr, acronym, address, big, cite, code,
-del, dfn, em, img, ins, kbd, q, s, samp,
-small, strike, strong, sub, sup, tt, var,
-b, u, i, center,
-dl, dt, dd, menu, ol, ul, li,
-fieldset, form, label, legend,
-table, caption, tbody, tfoot, thead, tr, th, td,
-article, aside, canvas, details, embed,
-figure, figcaption, footer, header, hgroup,
-main, menu, nav, output, ruby, section, summary,
-time, mark, audio, video {
-  margin: 0;
-  padding: 0;
-  border: 0;
-  font-size: 100%;
-  font: inherit;
-  vertical-align: baseline;
-}
-/* HTML5 display-role reset for older browsers */
-article, aside, details, figcaption, figure,
-footer, header, hgroup, main, menu, nav, section {
-  display: block;
-}
-/* HTML5 hidden-attribute fix for newer browsers */
-*[hidden] {
-    display: none;
-}
-body {
-  line-height: 1;
-}
-menu, ol, ul {
-  list-style: none;
-}
-blockquote, q {
-  quotes: none;
-}
-blockquote:before, blockquote:after,
-q:before, q:after {
-  content: '';
-  content: none;
-}
-table {
-  border-collapse: collapse;
-  border-spacing: 0;
-}
-*{
-	box-sizing: border-box;
-}
-body {
-  /* font-size: 100px; */
-  width: 100vw;
-  height: 100vh;
-	font-family: "Noto Sans KR", sans-serif;
-  background-color: ${(props) => props.theme.bgColor};
-  color: ${(props) => props.theme.textColor};
-  line-height: 1.4;
-  transition: 0.5s ease-in-out all;
-}
-a {
-	text-decoration: none;
-  color: inherit;
-}
-`;
+import { useRecoilState, useRecoilValue } from "recoil";
+import { hourSelector, minuteState } from "./atoms";
+import React from "react";
 
 function App() {
+	const [minutes, setMinutes] = useRecoilState(minuteState);
+	const [hours, setHours] = useRecoilState(hourSelector);
+	const onMinutesChange = (event: React.FormEvent<HTMLInputElement>) => {
+		setMinutes(+event.currentTarget.value);
+	};
+	const onHoursChange = (event: React.FormEvent<HTMLInputElement>) => {
+		setHours(+event.currentTarget.value);
+	};
 	return (
-		<>
-			<GlobalStyle />
-			<ToDoList />
-		</>
+		<div>
+			<input value={minutes} onChange={onMinutesChange} type="number" placeholder="Minute" />
+			<input value={hours} onChange={onHoursChange} type="number" placeholder="Hour" />
+		</div>
 	);
 }
