+++ src/atoms.ts
@@ -1,32 +1,18 @@
 import { atom, selector } from "recoil";
 
-export const enum Categories {
-	"TO_DO" = "TO_DO",
-	"DOING" = "DOING",
-	"DONE" = "DONE",
-}
-
-export interface IToDo {
-	text: string;
-	id: number;
-	category: Categories;
-}
-
-export const categoryState = atom<Categories>({
-	key: "category",
-	default: Categories.TO_DO,
+export const minuteState = atom({
+	key: "minute",
+	default: 0,
 });
 
-export const toDoState = atom<IToDo[]>({
-	key: "toDo",
-	default: [],
-});
-
-export const toDoSelector = selector({
-	key: "toDoSelector",
+export const hourSelector = selector({
+	key: "hours",
 	get: ({ get }) => {
-		const toDos = get(toDoState);
-		const category = get(categoryState);
-		return toDos.filter((toDo) => toDo.category === category);
+		const minutes = get(minuteState);
+		return minutes / 60;
+	},
+	set: ({ set }, newValue) => {
+		const minutes = +newValue * 60;
+		set(minuteState, minutes);
 	},
 });
